:art:add: 유저 정보 api LoginPage.js -> HomePage.js 에서 호출
